feat: move textVariants and textStyles to typography
