+++ lib/components/Text/index.ts
@@ -1,4 +1,1 @@
-export { Text, type TextProps } from './Text';
-export { useTextStyles } from './useTextStyles';
-
-export type { TextStyleProps } from './useTextStyles';
+export * from './Text';
